MODULE_4/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
importlib-metadata==6.6.0
Markdown==3.4.3
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.5.0
urllib3==2.0.2
zipp==3.15.0
